+++ GameServer/keeps/Managers/Template Manager.cs
@@ -227,7 +227,7 @@ namespace DOL.GS.Keeps
 						{
 							guard.Model = MidgardHastener;
 							guard.Size = 50;
-							guard.Flags ^= (uint)GameNPC.eFlags.GHOST;
+							guard.Flags ^= GameNPC.eFlags.GHOST;
 							break;
 						}
 					case eRealm.Hibernia:
